third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger/properties-special.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/console-test.js"></script>
<script src="../../../inspector/elements-test.js"></script>
<script>

function onload()
{
    console.dir(Object(true));
    console.dir((function(a,b) { return a + b; }));
    console.dir((function(a,b) { return a + b; }).bind({}, 2));
    console.dir((function*() { yield [1,2,3] }));
    runTest();
}

function test() {
  ConsoleTestRunner.expandConsoleMessages(dumpConsoleMessages);

  function dumpConsoleMessages() {
    ConsoleTestRunner.dumpConsoleMessages(false, false, TestRunner.textContentWithLineBreaks);
    TestRunner.completeTest();
  }
}

</script>
</head>
<body onload="onload()">
<p>Tests how debugger presents special properties of closures, bound functions and object wrappers.</p>
</body>
</html>
